server/whatsamyth-backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# Bot frameworks
python-telegram-bot==20.8
discord.py==2.3.2

# ML/Embeddings
sentence-transformers==2.3.1
faiss-cpu==1.13.0
transformers==4.37.2
torch==2.9.1

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.3
httpx==0.26.0
duckduckgo-search==4.4.3

# Scheduling
apscheduler==3.10.4

# Graph
networkx==3.2.1

# TTS
TTS==0.22.0
pyttsx3==2.90

# Utilities
python-dotenv==1.0.1
numpy==1.26.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
